requirements.txt
python-dotenv
aiogram
scipy
matplotlib